+++ src/main/java/sayTheSpire/Output.java
@@ -28,7 +28,7 @@ import sayTheSpire.buffers.*;
 
 public class Output {
 
-    public static String modVersion = "Beta 0.2.0";
+    public static String modVersion = "Beta 0.2.1";
 
     public enum Direction {
         NONE, UP, DOWN, LEFT, RIGHT