+++ src/main/java/sayTheSpire/UIPatches/rewards/RewardItemPatch.java
@@ -1,5 +1,6 @@
 import com.evacipated.cardcrawl.modthespire.lib.SpirePatch;
 import com.megacrit.cardcrawl.rewards.RewardItem;
+import sayTheSpire.ui.PotionElement;
 import sayTheSpire.Output;
 
 @SpirePatch(clz = RewardItem.class, method = "update")
@@ -10,7 +11,8 @@ public class RewardItemPatch {
             Output.text(__instance.text, true);
             switch (__instance.type) {
             case POTION:
-                Output.setupBuffers(__instance.potion);
+                if (__instance.potion != null)
+                    Output.setUI(new PotionElement(__instance.potion, PotionElement.PotionLocation.COMBAT_REWARDS));
                 break;
             case RELIC:
                 Output.setupBuffers(__instance.relic);