+++ src/main/java/sayTheSpire/UIPatches/ui/panels/TopPanelPatch.java
@@ -3,6 +3,7 @@ import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.potions.AbstractPotion;
 import com.megacrit.cardcrawl.potions.PotionSlot;
 import com.megacrit.cardcrawl.ui.panels.TopPanel;
+import sayTheSpire.ui.PotionElement;
 import sayTheSpire.Output;
 
 @SpirePatch(clz = TopPanel.class, method = "updatePotions")
@@ -11,12 +12,8 @@ public class TopPanelPatch {
     public static void Postfix(TopPanel __instance) {
         for (AbstractPotion p : AbstractDungeon.player.potions) {
             if (p.hb.justHovered) {
-                if (p instanceof PotionSlot) {
-                    Output.text("Empty potion slot", true);
-                } else {
-                    Output.text(p.name, true);
-                }
-                Output.setupBuffers(p);
+                PotionElement potionElement = new PotionElement(p, PotionElement.PotionLocation.MAIN_SCREEN);
+                Output.setUI(potionElement);
             }
         }
     }